Refactor JavaScript into modules and secure Firebase credentials.
This commit refactors the monolithic `script.js` file into a modular structure, separating concerns into `api.js`, `auth.js`, `firebase.js`, `main.js`, and `ui.js`. This improves code organization, readability, and maintainability.

Additionally, the Firebase configuration has been moved to a separate `firebase-config.js` file and added to `.gitignore` to prevent committing sensitive credentials to version control.
