+++ public/firebase.js
@@ -1,0 +1,6 @@
+// --- INITIALISATION FIREBASE ---
+firebase.initializeApp(firebaseConfig);
+const auth = firebase.auth();
+const db = firebase.firestore();
+
+export { auth, db };
